requirements.txt
fastapi
httpx
uvicorn
python-dotenv
jinja2==3.1.5
